% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{crearGato}
\alias{crearGato}
\title{Crea un dibujo de un gato en ascii}
\usage{
crearGato(numeroDeGatos)
}
\arguments{
\item{numeroDeGatos}{el numero de gatos a imprimir}
}
\value{
none
}
\description{
Crea un dibujo de un gato en ascii
}
\examples{

crearGato(10)

}
