% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.r
\docType{data}
\name{datosPruebaShannon}
\alias{datosPruebaShannon}
\title{Este es un set de datos creado como una distribuci贸n de Poisson con media 10 para probar la funci贸n shannon}
\format{LaA data frame with n rows and m variables:
        blah blah blah}
\source{
\url{http://github/lsaravia}
}
\usage{
datosPruebaShannon
}
\description{
Este es un set de datos creado como una distribuci贸n de Poisson con media 10 para probar la funci贸n shannon
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{datasets}
