% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fixClorophylData}
\alias{fixClorophylData}
\title{Fix the problems of chlorophyll-a data set}
\usage{
fixClorophylData(chla)
}
\arguments{
\item{chla}{data.frame with original data}
}
\value{
data.frame with fixed data
}
\description{
Fix the problems of chlorophyll-a data set
}
\examples{

fixClorophylData(chla)
}
