% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{shannon}
\alias{shannon}
\title{Calculates the Shannon Diversity Index using ln base e}
\usage{
shannon(species)
}
\arguments{
\item{species}{vector of abundances of species}
}
\value{
the Shannon Diversity Index
}
\description{
Calculates the Shannon Diversity Index using ln base e
}
\examples{

shannon(datosPruebaShannon)

}
